ref(debuginfo): Restructure variable and dwarf modules
