+++ symbolic-debuginfo/src/base/variable.rs
