+++ symbolic-debuginfo/src/lib.rs
@@ -36,7 +36,6 @@
 #![warn(missing_docs)]
 
 mod base;
-mod variable;
 
 #[cfg(all(
     feature = "breakpad",
@@ -75,7 +74,6 @@ pub mod sourcebundle;
 pub mod wasm;
 
 pub use crate::base::*;
-pub use crate::variable::*;
 
 #[cfg(all(
     feature = "breakpad",
